Extract core types to language_model_core
