+++ crates/agent_ui/src/agent_model_selector.rs
@@ -88,9 +88,9 @@ impl AgentModelSelector {
 impl Render for AgentModelSelector {
     fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
         let model = self.selector.read(cx).delegate.active_model(cx);
-        let model_name = model
+        let model_name: SharedString = model
             .as_ref()
-            .map(|model| model.model.name().0)
+            .map(|model| model.model.name().0.to_string().into())
             .unwrap_or_else(|| SharedString::from("Select a Model"));
 
         let provider_icon = model.as_ref().map(|model| model.provider.icon());
