+++ crates/language_model_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "open_router"
+name = "language_model_core"
 version = "0.1.0"
 edition.workspace = true
 publish.workspace = true
@@ -9,20 +9,19 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/open_router.rs"
-
-[features]
-default = []
-schemars = ["dep:schemars"]
+path = "src/language_model_core.rs"
+doctest = false
 
 [dependencies]
 anyhow.workspace = true
+cloud_llm_client.workspace = true
+derive_more.workspace = true
 futures.workspace = true
+gpui_util.workspace = true
 http_client.workspace = true
-language_model_core.workspace = true
-schemars = { workspace = true, optional = true }
+partial-json-fixer.workspace = true
+schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-settings.workspace = true
-strum.workspace = true
+smol.workspace = true
 thiserror.workspace = true
